Create final output
The code works on the user-defined function called myMean. Here in this program the data is given in a variable called assignment c(16, 18, 14, 22, 27, 17, 19, 17, 17, 22, 20, 22). A variable named assn is created and all the given data is been assigned to variable "assn". The logic is been set for the user-defined function called myMean where it finds the mean of the given data with mathematical formulation (sum(assn)/length(assn)). The program has been executed and the output was printed correctly since the logic given function myMean is correct and it works accordingly.